move watermark shortcuts closer to letterpress
fixes #59952
