+++ src/vs/workbench/browser/parts/editor/media/editorgroupview.css
@@ -18,27 +18,99 @@
 	opacity: 1; /* indicate active/dragged-over group through undimmed state */
 }
 
-/* Letterpress */
+/* Watermark & shortcuts */
 
-.monaco-workbench .part.editor > .content .editor-group-container > .editor-group-letterpress  {
-	display: none; /* only visible when empty */
-}
-
-.monaco-workbench .part.editor > .content .editor-group-container.empty > .editor-group-letterpress  {
-	display: block;
+.monaco-workbench .part.editor > .content .editor-group-container > .editor-group-watermark  {
+	display: flex;
+	height: 100%;
+	max-width: 290px;
 	margin: auto;
-	width: 100%;
-	height: calc(100% - 70px); /* centered below toolbar */
-	max-width: 260px;
-	background-repeat: no-repeat;
-	background-position: 50% 50%;
-	background-size: 70% 70%;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
 }
 
-.monaco-workbench .part.editor > .content.empty .editor-group-container.empty > .editor-group-letterpress  {
-	background-size: 100% 100%; /* larger for empty editor part */
-	height: 100%; /* no toolbar in this case */
+.monaco-workbench .part.editor > .content .editor-group-container:not(.empty) > .editor-group-watermark {
+	display: none;
 }
+
+.monaco-workbench .part.editor > .content:not(.empty) .editor-group-container.empty > .editor-group-watermark {
+	max-width: 200px;
+	height: calc(100% - 35px);
+}
+
+.monaco-workbench .part.editor > .content .editor-group-container > .editor-group-watermark > .letterpress {
+	width: 100%;
+	aspect-ratio: 1/1;
+	background-image: url('./letterpress-light.svg');
+	background-position-x: center;
+	background-repeat: no-repeat;
+}
+
+.monaco-workbench.vs-dark .part.editor > .content .editor-group-container .editor-group-watermark > .letterpress {
+	background-image: url('./letterpress-dark.svg');
+}
+
+.monaco-workbench.hc-light .part.editor > .content .editor-group-container .editor-group-watermark > .letterpress {
+	background-image: url('./letterpress-hcLight.svg');
+}
+
+.monaco-workbench.hc-black .part.editor > .content .editor-group-container .editor-group-watermark > .letterpress {
+	background-image: url('./letterpress-hcDark.svg');
+}
+
+.monaco-workbench .part.editor > .content:not(.empty) .editor-group-container > .editor-group-watermark > .shortcuts,
+.monaco-workbench .part.editor > .content .editor-group-container.max-height-478px > .editor-group-watermark > .shortcuts {
+	display: none;
+}
+
+.monaco-workbench .part.editor > .content .editor-group-container > .editor-group-watermark > .shortcuts > .watermark-box {
+	display: inline-table;
+	border-collapse: separate;
+	border-spacing: 11px 17px;
+}
+
+.monaco-workbench .part.editor > .content .editor-group-container > .editor-group-watermark > .shortcuts dl {
+	display: table-row;
+	opacity: .8;
+	cursor: default;
+}
+
+.monaco-workbench .part.editor > .content .editor-group-container > .editor-group-watermark > .shortcuts dt {
+	text-align: right;
+	letter-spacing: 0.04em
+}
+
+.monaco-workbench .part.editor > .content .editor-group-container > .editor-group-watermark > .shortcuts dd {
+	text-align: left;
+}
+
+.monaco-workbench .part.editor > .content .editor-group-container > .editor-group-watermark > .shortcuts dt,
+.monaco-workbench .part.editor > .content .editor-group-container > .editor-group-watermark > .shortcuts dd {
+	display: table-cell;
+	vertical-align: middle;
+}
+
+.monaco-workbench .part.editor > .content .editor-group-container > .editor-group-watermark > .shortcuts dt,
+.monaco-workbench .part.editor > .content .editor-group-container > .editor-group-watermark > .shortcuts dl {
+	color: rgba(0,0,0,.68);
+}
+
+.monaco-workbench.vs-dark .part.editor > .content .editor-group-container .editor-group-watermark > .shortcuts dt,
+.monaco-workbench.vs-dark .part.editor > .content .editor-group-container .editor-group-watermark > .shortcuts dl {
+	color: rgba(255,255,255,.6);
+}
+
+.monaco-workbench.hc-black .part.editor > .content .editor-group-container .editor-group-watermark > .shortcuts dt,
+.monaco-workbench.hc-light .part.editor > .content .editor-group-container .editor-group-watermark > .shortcuts dt {
+	color: var(--vscode-editor-foreground);
+}
+.monaco-workbench.hc-black .part.editor > .content .editor-group-container .editor-group-watermark > .shortcuts dl,
+.monaco-workbench.hc-light .part.editor > .content .editor-group-container .editor-group-watermark > .shortcuts dl {
+	color: var(--vscode-editor-foreground);
+	opacity: 1;
+}
+
 
 /* Title */
