--- src/vs/workbench/contrib/watermark/browser/media/watermark.css
@@ -1,76 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-.monaco-workbench .part.editor.has-watermark > .content.empty .editor-group-container > .editor-group-letterpress {
-	background-position-y: calc(50% - 100px);
-}
-
-.monaco-workbench .part.editor.max-height-478px > .content.empty .editor-group-container > .editor-group-letterpress {
-	background-position-y: 50%;
-}
-
-.monaco-workbench .part.editor > .content > .watermark {
-	display: none; /* only visible when no editors are opened */
-}
-
-.monaco-workbench .part.editor.has-watermark > .content.empty > .watermark {
-	display: block;
-	position: absolute;
-	width: 100%;
-	top: calc(50% + 55px);
-	text-align: center;
-	white-space: nowrap;
-	overflow: hidden;
-}
-
-.monaco-workbench .part.editor > .content.empty > .watermark > .watermark-box {
-	display: inline-table;
-	border-collapse: separate;
-	border-spacing: 13px 17px;
-}
-
-.monaco-workbench .part.editor.max-height-478px > .content.empty > .watermark {
-	display: none;
-}
-
-.monaco-workbench .part.editor > .content.empty > .watermark dl {
-	display: table-row;
-	opacity: .8;
-	cursor: default;
-}
-
-.monaco-workbench .part.editor > .content.empty > .watermark dt {
-	text-align: right;
-	letter-spacing: 0.04em
-}
-
-.monaco-workbench .part.editor > .content.empty > .watermark dd {
-	text-align: left;
-}
-
-.monaco-workbench .part.editor > .content.empty > .watermark dt,
-.monaco-workbench .part.editor > .content.empty > .watermark dd {
-	display: table-cell;
-}
-
-.monaco-workbench .part.editor > .content.empty > .watermark dt,
-.monaco-workbench .part.editor > .content.empty > .watermark dl {
-	color: rgba(0,0,0,.68);
-}
-
-.monaco-workbench.vs-dark .part.editor > .content.empty > .watermark dt,
-.monaco-workbench.vs-dark .part.editor > .content.empty > .watermark dl {
-	color: rgba(255,255,255,.6);
-}
-
-.monaco-workbench.hc-black .part.editor > .content.empty > .watermark dt,
-.monaco-workbench.hc-light .part.editor > .content.empty > .watermark dt {
-	color: var(--vscode-editor-foreground);
-}
-.monaco-workbench.hc-black .part.editor > .content.empty > .watermark dl,
-.monaco-workbench.hc-light .part.editor > .content.empty > .watermark dl {
-	color: var(--vscode-editor-foreground);
-	opacity: 1;
-}
